Patrol state, distance and target functions
Made the state where the ai searches for the target. Distance functions for checking if they're within the distance of the target and if they're the closest tank (and to be the target if so). Lastly, functions for checking if there's a target, and for automatically assigning the first player as the the target.