Ai Vision, Distance Manipulation, Debugging
Debugged the flee and chase state, Also, the hasTarget() function. Implemented a distance multiplier that increases the flee/chase distance upon the start of the state. The increase is based on the distance multiplier float. After, the distance has been reached it returns to normal(this is just to make the actions seem less stiff and sudden). Began working on the ai vision and hearing.