Heal overtime and Gamblehealth powerups, instantDamage powerup
Implemented the heal overtime function that heals based on however often specified, the heal amount, and the duration to heal. The gamble health power up takes the gambled amount from the current health, then when the power up is removed they gain the gambled amount * the gamble multiplier. Also, the instant damage power up was added and made a prefab(along with the rest of the power ups).