New main scene, Game states, Score
Made a new main scene that contains various game states such as title, game, etc. The game states can be switched between based on the function that's called that'll disable the rest of the states. The score can be added to or subtracted from via power ups or killing a pawn.