Timer, Win Screen, Gameobject wipe
Added a timer that counts down when the game is launched. When this timer reaches zero, the game goes to game over. Made the win screen Ui and state(along with the code). Lastly, I made a list that stores all the pawns and purges them when the game ends because sometimes it wouldn't.